Pack 3 TT entries in 32 bytes cluster
Idea from Ron Britvich

Code reworked by Marco Costalba and Joona Kiiski

Bench: 8095369

Resolves #3
Resolves #10

Conflicts:

	src/tt.cpp
	src/tt.h

+++ src/tt.h
@@ -94,7 +94,7 @@ public:
   void store(const Key key, Value v, Bound type, Depth d, Move m, Value statV);
 
 private:
-  size_t clusterCount;
+  uint32_t clusterCount;
   TTCluster* table;
   void* mem;
   uint8_t generation; // Size must be not bigger than TTEntry::genBound8
